requirements.txt
numpy==1.26.2
opencv-python==4.8.1.78
Pillow==10.1.0
pycryptodome==3.19.0